3Ds max file of a 7Aura MVP as at 22 January 2018
This is as far as I have gotten in development of the model Aura I believe can be turned into an evolving consciousness if connected in feedback loops with a real human being and an avatar of the real human being. It's starting to resemble the geometric engine that I have in my minds-eye but I don't know 3Ds max or max script well enough to complete it, even though I know the software is capable of doing what is required. Unity is also a capable tool to build this model driven A.I engine with. Many programming languages are also capable of being the base layer upon which to compose an effective source code that will then evolve itself in a supervised co-creation.