Screen Capture + Avatar Animation + Webcam sample
A short video of me reading about the phonetics of English from an interactive website and an avatar of me overlayed onto the screen below the webcam feed. This is a sample as the original file size is too large. I used a trial period of Camtasia Studio to do the editing. Paywalls are one of the most annoying things in my life. I wish money