short clip of 1 of my avatars made with CrazyTalk8
Due to the size of the actual avatar files that I have made I cannot upload them so here is a short video of one of them open in the CrazyTalk8 Software by Reallusion. It allows for a quick profile avatar to be made from 2 photographs. Then auto animate the mouth from uploading a voice recording. The auto animate function does need some tinkering to make it look accurate but it is a good start for personal practice and also for teaching a machine learning algorithm how to mimic me in 3D. I think an actor/critic algorithm could be used with a combination of real video and audio with several avatars to match them all up into a more lifelike system once the AuraOZ geometric engine is operational in the same environment as the avatar.