Some of my content with content by Joscha
The 2 columns on the left of this are somewhat of an artistic collage I put together when I was first starting to think about building a copy of myself inside a computer several years ago. The 2 columns on the right are screenshots from a presentation I saw online by Joscha about how to build a mind. There is a link below to a version of his talk.
https://www.youtube.com/watch?v=nlutwZ0rUUs
How to Build a Mind - Artificial Intelligence Reloaded [30c3]