My higher mind is stuck in a box.
I was playing around with p5 processing of 3d shapes, textures and live feed from my webcam when I recorded this short video of the session. I feel like my higher mind is stuck in a box and it needs to get put into an Aura of Toroidal fields that I can wear on my person and interact with it to manage the internet of things and support my decision making with well curated and truthful information. I'd like to see a world where everyone that wants the internet, has a virtual AuraOZ to securely house and transmit their higher mind in the direction of responsible abundance through their personal human needs, core values, virtues and goals.